refactor: separate profile picture fetching logic into useImage hook and update AccountPage to utilize it
